scripts/requirements.txt
can-isotp==2.0.3 # uds sample and runner
udsoncan==1.21.2 # uds sample and runner
python-can==4.6.1 # can log receiver
